+++ esp-hal/src/soc/esp32/cpu_control.rs
@@ -1,13 +1,11 @@
 //! # Control CPU Cores (ESP32)
 //!
 //! ## Overview
-//!
 //! This module provides essential functionality for controlling
 //! and managing the APP (second) CPU core on the `ESP32` chip. It is used to
 //! start and stop program execution on the APP core.
 //!
-//! ## Example
-//!
+//! ## Examples
 //! ```rust, no_run
 #![doc = crate::before_snippet!()]
 //! # use esp_hal::delay::Delay;
@@ -82,6 +80,12 @@ use crate::{
 pub struct Stack<const SIZE: usize> {
     /// Memory to be used for the stack
     pub mem: MaybeUninit<[u8; SIZE]>,
+}
+
+impl<const SIZE: usize> Default for Stack<SIZE> {
+    fn default() -> Self {
+        Self::new()
+    }
 }
 
 #[allow(clippy::len_without_is_empty)]