+++ esp-hal/src/spi/master.rs
@@ -1,12 +1,36 @@
 //! # Serial Peripheral Interface - Master Mode
 //!
 //! ## Overview
+//! In this mode, the SPI acts as master and initiates the SPI transactions.
 //!
-//! There are multiple ways to use SPI, depending on your needs. Regardless of
-//! which way you choose, you must first create an SPI instance with
-//! [`Spi::new`].
+//! ## Configuration
+//! The peripheral can be used in full-duplex and half-duplex mode and can
+//! leverage DMA for data transfers. It can also be used in blocking or async.
+//!
+//! ### Exclusive access to the SPI bus
+//! If all you want to do is to communicate to a single device, and you initiate
+//! transactions yourself, there are a number of ways to achieve this:
+//!
+//! - Use the [`FullDuplex`](embedded_hal_02::spi::FullDuplex) trait to
+//!   read/write single bytes at a time,
+//! - Use the [`SpiBus`](embedded_hal::spi::SpiBus) trait (requires the
+//!   "embedded-hal" feature) and its associated functions to initiate
+//!   transactions with simultaneous reads and writes, or
+//! - Use the `ExclusiveDevice` struct from [`embedded-hal-bus`] or `SpiDevice`
+//!   from [`embassy-embedded-hal`].
+//!
+//!
+//! ### Shared SPI access
+//! If you have multiple devices on the same SPI bus that each have their own CS
+//! line, you may want to have a look at the implementations provided by
+//! [`embedded-hal-bus`] and [`embassy-embedded-hal`].
+//!
+//! ## Usage
+//! The module implements several third-party traits from embedded-hal@0.2.x,
+//! embedded-hal@1.x.x and embassy-embedded-hal
 //!
 //! ## Example
+//! ### SPI Initialization
 //! ```rust, no_run
 #![doc = crate::before_snippet!()]
 //! # use crate::esp_hal::prelude::_fugit_RateExtU32;
@@ -29,26 +53,6 @@
 //! # }
 //! ```
 //! 
-//! ## Exclusive access to the SPI bus
-//!
-//! If all you want to do is to communicate to a single device, and you initiate
-//! transactions yourself, there are a number of ways to achieve this:
-//!
-//! - Use the [`FullDuplex`](embedded_hal_02::spi::FullDuplex) trait to
-//!   read/write single bytes at a time,
-//! - Use the [`SpiBus`](embedded_hal::spi::SpiBus) trait (requires the
-//!   "embedded-hal" feature) and its associated functions to initiate
-//!   transactions with simultaneous reads and writes, or
-//! - Use the `ExclusiveDevice` struct from [`embedded-hal-bus`] or `SpiDevice`
-//!   from [`embassy-embedded-hal`].
-//!
-//!
-//! ## Shared SPI access
-//!
-//! If you have multiple devices on the same SPI bus that each have their own CS
-//! line, you may want to have a look at the implementations provided by
-//! [`embedded-hal-bus`] and [`embassy-embedded-hal`].
-//!
 //! [`embedded-hal-bus`]: https://docs.rs/embedded-hal-bus/latest/embedded_hal_bus/spi/index.html
 //! [`embassy-embedded-hal`]: https://docs.embassy.dev/embassy-embedded-hal/git/default/shared_bus/index.html