+++ esp-hal/src/timer/timg.rs
@@ -1,12 +1,16 @@
 //! # Timer Group (TIMG)
 //!
+//! ## Overview
 //! The Timer Group (TIMG) peripherals contain one or more general-purpose
 //! timers, plus one or more watchdog timers.
 //!
 //! The general-purpose timers are based on a 16-bit pre-scaler and a 54-bit
-//! auto-reload-capable up-down counter. The timers have configurable alarms,
-//! which are triggered when the internal counter of the timers reaches a
-//! specific target value. The timers are clocked using the APB clock source.
+//! auto-reload-capable up-down counter.
+//!
+//! ## Configuration
+//! The timers have configurable alarms, which are triggered when the internal
+//! counter of the timers reaches a specific target value. The timers are
+//! clocked using the APB clock source.
 //!
 //! Typically, a general-purpose timer can be used in scenarios such as:
 //!
@@ -14,12 +18,9 @@
 //! - Generate one-shot alarms; trigger events once
 //! - Free-running; fetching a high-resolution timestamp on demand
 //!
-//! ## Usage
 //!
-//! ### Examples
-//!
-//! #### General-purpose Timer
-//!
+//! ## Examples
+//! ### General-purpose Timer
 //! ```rust, no_run
 #![doc = crate::before_snippet!()]
 //! # use esp_hal::timer::timg::TimerGroup;
@@ -517,15 +518,40 @@ where
     }
 
     fn enable_interrupt(&self, state: bool) {
+        // always use level interrupt
+        #[cfg(any(esp32, esp32s2))]
+        self.register_block()
+            .t(self.timer_number().into())
+            .config()
+            .modify(|_, w| w.level_int_en().set_bit());
+
         self.register_block()
             .int_ena_timers()
-            .write(|w| w.t(self.timer_number()).bit(state));
+            .modify(|_, w| w.t(self.timer_number()).bit(state));
     }
 
     fn clear_interrupt(&self) {
         self.register_block()
             .int_clr_timers()
             .write(|w| w.t(self.timer_number()).clear_bit_by_one());
+    }
+
+    fn set_interrupt_handler(&self, handler: InterruptHandler) {
+        let interrupt = match (self.timer_group(), self.timer_number()) {
+            (0, 0) => Interrupt::TG0_T0_LEVEL,
+            #[cfg(timg_timer1)]
+            (0, 1) => Interrupt::TG0_T1_LEVEL,
+            #[cfg(timg1)]
+            (1, 0) => Interrupt::TG1_T0_LEVEL,
+            #[cfg(all(timg_timer1, timg1))]
+            (1, 1) => Interrupt::TG1_T1_LEVEL,
+            _ => unreachable!(),
+        };
+
+        unsafe {
+            interrupt::bind_interrupt(interrupt, handler.handler());
+        }
+        interrupt::enable(interrupt, handler.priority()).unwrap();
     }
 
     fn is_interrupt_set(&self) -> bool {
@@ -547,6 +573,8 @@ where
 #[doc(hidden)]
 pub trait Instance: Sealed + Enable {
     fn register_block(&self) -> &RegisterBlock;
+
+    fn timer_group(&self) -> u8;
 
     fn timer_number(&self) -> u8;
 
@@ -630,6 +658,10 @@ where
 {
     fn register_block(&self) -> &RegisterBlock {
         unsafe { &*TG::register_block() }
+    }
+
+    fn timer_group(&self) -> u8 {
+        TG::id()
     }
 
     fn timer_number(&self) -> u8 {